pro mode: snap shutter slider to a ½-stop standard ladder
Was a continuous log-scale slider rounded to whole ms, which produced
non-standard speeds (1/58 instead of 1/60), jittered as the value binding
fed back mid-drag, and crammed fast speeds into the left edge.  Now the
slider steps a fixed ½-stop ladder (1/1000 … 1/3) plus the HAL's max-exposure
cap as the final rung: travel is even per stop, every value is a conventional
speed, and the label reads exactly.

Two precision fixes make the speeds land true:
- pass fractional ms through setManualExposure (1/60 = 16.667ms; rounding to
  whole ms sent 17ms and the EXIF read 1/58)
- floor the advertised exposure minimum at 1ms — the sub-ms HAL minimum
  truncated to 0 and left the slider's bottom reading '1/Infinity'

The ladder is device-independent, so a phone advertising a longer max
exposure simply gets more rungs at the same ½-stop spacing.

+++ src/camera2/Camera2Bridge.h
@@ -229,7 +229,7 @@ public:
     //   setManualExposure(iso, exposureMs) — switch to AE_MODE_OFF and apply values
     //   setExposureCompensation(ev)        — 0.0 most under, 0.5 neutral, 1.0 most over
     Q_INVOKABLE void setAutoExposure();
-    Q_INVOKABLE void setManualExposure(int iso, int exposureMs);
+    Q_INVOKABLE void setManualExposure(int iso, double exposureMs);
     Q_INVOKABLE void setExposureCompensation(float ev);
     Q_INVOKABLE void setFocusDistance(float diopters);   // 0=infinity
